examples/python/requirements.txt
evalai-sdk>=1.0.0
python-dotenv>=1.0.0
